feature-planningRoutes
- The primitive type of xAxe and yAxe in CartesianCoordinate, have been change to int
- The class FileManagerToRoutePlanningCmd has been created
- The class RoutePlanningToFileManagerDtoFailException has been created
- The class PlanningRoutesToDroneManagerCmd has been renamed to RoutesPlanningToDroneManagerCmd
- The class CircularValueAndCoordinate has been created
- The class ValueAndCoordinatehas been created
- The enum CoordinatesDirection has been created
- The Exception BusinessException has been created
- The class RoutePlanningImpl has been created
- The interface RoutePlanning has been created
- The class RoutePlanningDto has been created

+++ SuCorrientazoADomicilioMarlonR/src/main/java/s4n/codechallenge/actorsdtos/commands/CartesianCoordinateCmd.java
@@ -11,8 +11,8 @@ import s4n.codechallenge.entities.CartesianCoordinate;
 @Setter
 @Builder
 public class CartesianCoordinateCmd {
-    private byte xAxe;
-    private byte yAxe;
+    private int xAxe;
+    private int yAxe;
 
     public static CartesianCoordinateCmd toCmd(CartesianCoordinate cartesianCoordinate) {
         return CartesianCoordinateCmd.builder()
